requirements.txt
tensorflow
streamlit
numpy
pillow
matplotlib
h5py
